Добавлено размещение имитаторов КПИУ в одном виджете, определены ограничения размеров окон виждетов, поправлены типы параметров функций имитаторов.

+++ UNFOI/rpc_foi.cpp
@@ -41,7 +41,7 @@ RpcFoiWidget::RpcFoiWidget() : QWidget()
 	}
 }
 
-int RpcFoiWidget::unfoi_map_channels_setup(int _n, short _chan)
+int RpcFoiWidget::unfoi_map_channels_setup(int _n, int _chan)
 {
 	map_channels[_n] = _chan;
 	SRPCSignalClass::Instance().toLog(QString("Присваиваю каналу %1 значение %2").arg(_n).arg(_chan));
